Add notification repository and integrate in dashboard
